add logger

+++ data-features/feature-api/src/commonMain/kotlin/com/grippo/data.features.api/user/models/User.kt
@@ -8,5 +8,5 @@ public data class User(
     val email: String,
     val experience: ExperienceEnumEnum,
     val weight: Float,
-    val height: Float
+    val height: Int
 )
